libs/nx-python/src/generators/grpc-project/files/base/pyproject.toml
[tool.poetry]
name = "<%= projectName %>"
version = "0.0.0"
description = "<%= description %>"
authors = []
readme = 'README.md'
packages = [{ include = "<%= moduleName %>" }]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.54.2"
protobuf = "^4.23.4"
grpcio-reflection = "^1.54.2"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.5"
black = "^23.7.0"
pytest = "^7.4.0"
pyright = "^1.1.325"

[tool.poetry.scripts]
dev = "<%= moduleName %>.main:dev"

[tool.black]
line-length = 120
target-version = ["py310"]
workers = 4

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
